examples/data/requirements.txt
numpy
pandas
imageio
progressbar2